rename DataToEncodeModal to DataTemplatesModal, and add to README
